+++ src/main/java/net/neoforged/neoforge/internal/RegistrationEvents.java
@@ -10,7 +10,7 @@ import net.neoforged.neoforge.capabilities.CapabilityHooks;
 import net.neoforged.neoforge.common.world.chunk.ForcedChunkManager;
 import net.neoforged.neoforge.event.ModifyDefaultComponentsEvent;
 import net.neoforged.neoforge.fluids.CauldronFluidContent;
-import net.neoforged.neoforge.network.syncreg.RegistryConfigHandlers;
+import net.neoforged.neoforge.network.registryconfigsync.RegistryConfigHandlers;
 import net.neoforged.neoforge.registries.RegistryManager;
 
 public class RegistrationEvents {
